death to trade_item asm and unify into one C file
