--- asm/trade_items_menu.s
@@ -1,33 +1,0 @@
-        .include "constants/gba_constants.inc"
-	.include "asm/macros.inc"
-
-	.syntax unified
-
-	.text
-
-	thumb_func_start sub_80369D0
-sub_80369D0:
-	ldr r0, _080369F4
-	ldr r2, [r0]
-	ldr r0, _080369F8
-	ldr r1, [r0]
-	movs r3, 0x95
-	lsls r3, 2
-	adds r0, r1, r3
-	ldrb r0, [r0]
-	lsls r0, 1
-	adds r2, 0x50
-	adds r2, r0
-	ldrh r0, [r2]
-	adds r3, 0x4
-	adds r1, r3
-	ldr r1, [r1]
-	adds r0, r1
-	strh r0, [r2]
-	bx lr
-	.align 2, 0
-_080369F4: .4byte gUnknown_203B460
-_080369F8: .4byte gUnknown_203B358
-	thumb_func_end sub_80369D0
-
-        .align 2,0
