+++ ld_script.txt
@@ -104,8 +104,6 @@ SECTIONS {
         src/main_menu_mid.o(.text);
         asm/main_menu_1.o(.text);
         src/trade_items_menu.o(.text);
-        asm/trade_items_menu.o(.text);
-        src/trade_items_menu_1.o(.text);
         asm/other_menus.o(.text);
         asm/rescue_password_menu.o(.text);
         asm/load_screen.o(.text);
